+++ src/routes/routes.tsx
@@ -2,5 +2,5 @@ export const ROUTES = {
   HOME: '/',
   LOGIN: '/login',
   SIGNUP: '/signup',
-  TFPAGE: '/tfpage',
+  TFPAGE: '/otp',
 };
